refactoring and PlayControl handlers
