+++ src/components/App/App.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Sketch from '../Sketch/Sketch';
 import Header from '../Header/Header';
-import PlayControl from '../Synth/PlayControl';
+import PlayControl from '../Synth/PlayControl/PlayControl';
 
 export default function App() {
 
@@ -10,7 +10,7 @@ export default function App() {
     <>
       <Header />
       <Sketch />
-      <PlayControl />
+
     </>
   );
 }
