+++ src/fonts/segmentNoteMap.js
@@ -1,4 +1,4 @@
-import { scaleMap } from './scaleMap';
+import { scaleMap } from '../utils/scaleMap';
 export const getSegmentNoteMap = (scale) => {
   return scaleMap.scale;
 };
